job_scraper/pyproject.toml
[project]
name = "job_scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "beautifulsoup4>=4.14.2",
    "celery>=5.5.3",
    "fastapi[standard]>=0.120.4",
    "greenlet>=3.2.4",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-openai>=0.3.35",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
    "typer>=0.19.2",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.12.9",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-requests>=2.32.4.20250913",
]
